Move unnecessary classes for pypi package, include all necessary packages in setup.py
