+++ main/gbd/database/benchmark_administration.py
@@ -4,7 +4,7 @@ from os.path import isfile
 from main.gbd.database import groups
 from main.gbd.database.db import Database, DatabaseException
 from main.gbd.hashing.gbd_hash import gbd_hash
-from util import eprint
+from main.util.util import eprint
 
 
 def add_tag(database, cat, tag, hash, force=False):
