+++ main/gbd/database/db.py
@@ -2,7 +2,7 @@ import sqlite3
 from os.path import isfile
 
 from main.gbd.hashing.gbd_hash import HASH_VERSION
-from util import eprint
+from main.util.util import eprint
 
 VERSION = 0
